refactor: separate substrate listeners and endpoint.
