+++ src/endpoints/escrow/escrow.service.ts
@@ -1,21 +1,15 @@
 import { Injectable } from '@nestjs/common';
-import { formatEther, WalletSigner } from 'nestjs-ethers';
+import { WalletSigner } from 'nestjs-ethers';
 import { EthereumService } from '../ethereum/ethereum.service';
-import { SubstrateService } from '../../substrate/substrate.service';
-import { Utils } from './utils/utils';
+import { SubstrateService } from '../substrate/substrate.service';
 import { ethers } from 'ethers';
-import { ElasticsearchService } from '@nestjs/elasticsearch';
 
 @Injectable()
 export class EscrowService {
-  private utils: Utils;
   constructor(
     private substrateService: SubstrateService,
     private ethereumService: EthereumService,
-    private readonly elasticsearchService: ElasticsearchService,
-  ) {
-    this.utils = new Utils();
-  }
+  ) {}
 
   async createOrder(request) {
     console.log('[createOrder] request: ', request);
@@ -97,13 +91,4 @@ export class EscrowService {
       console.log(error);
     }
   }
-
-  /**
-   * Get refund gas estimation:
-   * Refund is done by calling the transferFrom function of ERC20 token
-   * The transfer is done from escrow's address to the customer address
-   * @param customerAddress : customer address
-   * @param amount : amount of erc20 being transferred to customer
-   * @returns string
-   */
 }
