+++ src/endpoints/substrate/dto/get-dbio-on-register.dto.ts
@@ -1,0 +1,11 @@
+import { ApiProperty } from "@nestjs/swagger";
+
+export type RegistrationRole = 'lab' | 'doctor' | 'hospital';
+
+export class GetDbioOnRegisterDto {
+  @ApiProperty({ type: String })
+  accountId: string;
+
+  @ApiProperty({ description: 'RegistrationRole' })
+  role: RegistrationRole;
+}
