+++ src/endpoints/substrate/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './get-dbio-on-register.dto';
+export * from './wallet-binding.dto';
