+++ src/endpoints/substrate/dto/wallet-binding.dto.ts
@@ -1,0 +1,9 @@
+import { ApiProperty } from "@nestjs/swagger";
+
+export class WalletBindingDTO {
+    @ApiProperty({ type: String })
+    accountId: string;
+  
+    @ApiProperty({ type: String })
+    ethAddress: string;
+}
