+++ src/endpoints/substrate/labs/lab.controller.ts
