+++ src/substrate/geneticTestingEvent.ts
@@ -2,7 +2,7 @@ import { Injectable, OnModuleInit } from '@nestjs/common';
 import { ApiPromise, WsProvider } from '@polkadot/api';
 import { RewardDto } from '../common/utilities/reward/dto/reward.dto';
 import { RewardService } from '../common/utilities/reward/reward.service';
-import { SubstrateService } from './substrate.service';
+import { SubstrateService } from '../endpoints/substrate/substrate.service';
 import { ethers } from 'ethers'
 import { DebioConversionService } from '../common/utilities/debio-conversion/debio-conversion.service';
