+++ src/substrate/orderEvent.ts
@@ -3,7 +3,7 @@ import { Option } from '@polkadot/types';
 import { EscrowService } from '../endpoints/escrow/escrow.service';
 import { RewardDto } from '../common/utilities/reward/dto/reward.dto';
 import { RewardService } from '../common/utilities/reward/reward.service';
-import { SubstrateService } from './substrate.service';
+import { SubstrateService } from '../endpoints/substrate/substrate.service';
 import { Logger } from '@nestjs/common';
 import { TransactionLoggingService } from '../common/utilities/transaction-logging/transaction-logging.service';
 import { TransactionLoggingDto } from '../common/utilities/transaction-logging/dto/transaction-logging.dto';
